feat: add room CR, UD left

+++ internal/model/model.go
@@ -12,8 +12,8 @@ type PageResponse[T any] struct {
 }
 
 type PageMetaData struct {
-	Page      int `json:"page"`
-	Size      int `json:"size"`
-	TotalItem int `json:"total_item"`
-	TotalPage int `json:"total_page"`
+	Page      int   `json:"page"`
+	Size      int   `json:"size"`
+	TotalItem int64 `json:"total_item"`
+	TotalPage int64 `json:"total_page"`
 }
